refactor: #14 이미지 관련 공통 서비스 클래스 생성
