+++ src/main/java/com/dnd/MusicLog/log/controller/LogController.java
@@ -4,8 +4,7 @@ import com.dnd.MusicLog.global.common.BaseController;
 import com.dnd.MusicLog.global.common.BaseResponse;
 import com.dnd.MusicLog.global.common.SuccessResponse;
 import com.dnd.MusicLog.global.jwt.util.JwtTokenProvider;
-import com.dnd.MusicLog.log.service.ImagesDeleteService;
-import com.dnd.MusicLog.log.service.ImagesUploadService;
+import com.dnd.MusicLog.log.service.ImagesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,14 +19,13 @@ import java.util.List;
 public class LogController extends BaseController {
 
     private final JwtTokenProvider jwtTokenProvider;
-    private final ImagesUploadService imagesUploadService;
-    private final ImagesDeleteService imagesDeleteService;
+    private final ImagesService imagesService;
 
     @PostMapping("/image")
     public ResponseEntity<BaseResponse<List<String>>> uploadImages(@RequestHeader(name = "Authorization") String bearerToken,
                                                                    @RequestPart("images") List<MultipartFile> multipartFile) {
         jwtTokenProvider.extractAccessTokenSubject(bearerToken);
-        List<String> responseDto = imagesUploadService.uploadImages(multipartFile, "images");
+        List<String> responseDto = imagesService.uploadImages(multipartFile, "images");
         return createBaseResponse(HttpStatus.CREATED, "이미지 저장 완료", responseDto);
     }
 
@@ -35,7 +33,7 @@ public class LogController extends BaseController {
     public ResponseEntity<SuccessResponse> deleteImage(@RequestHeader(name = "Authorization") String bearerToken,
                                                        @RequestParam String fileName) {
         jwtTokenProvider.extractAccessTokenSubject(bearerToken);
-        imagesDeleteService.deleteImage(fileName);
+        imagesService.deleteImage(fileName);
         return createSuccessResponse(HttpStatus.OK, "이미지 삭제 완료");
     }
