--- src/main/java/com/dnd/MusicLog/log/service/ImagesDeleteService.java
@@ -1,33 +1,0 @@
-package com.dnd.MusicLog.log.service;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
-import com.dnd.MusicLog.global.error.exception.BusinessLogicException;
-import com.dnd.MusicLog.global.error.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class ImagesDeleteService {
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
-    private final AmazonS3 amazonS3;
-
-    public void deleteImage(String fileName) {
-        getFileExtension(fileName);
-        amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
-    }
-
-    private String getFileExtension(String fileName) {
-        try {
-            return fileName.substring(fileName.lastIndexOf("."));
-        } catch (StringIndexOutOfBoundsException e) {
-            throw new BusinessLogicException(ErrorCode.BAD_REQUEST_FILENAME);
-        }
-    }
-
-}
